// src/app/core/constants/entities.enum.ts
export enum Entities {
  Sentences = 'sentencias',
  Sentences_pub = 'sentencias_pub',
  SentencesParagraph = 'sentencias_parrafos',
  SentencesParagraph_pub = 'sentencias_parrafos_pub',
  Thesis = 'tesis',
  Votes = 'votos',
  Legislation = 'legislacion',
  Library = 'biblioteca',
  Orderings = 'legislacion',
  Agreements = 'acuerdos',
  Protocols = 'protocolos_actuacion_dh',
  Versions = 'vtaquigraficas',
  Onu = 'suniversal',
  CIJ = 'cij',
  CoIDH = 'bjdh_coidh',
  CoIDH_BJ = 'bjdh',
  CIDH = 'cidh',
  Hudoc = 'hudoc',
  TSpanish = 'tcespanol',
  TColombia = 'cccolombia',
  TChile = 'tcchile',
  CSJNArgentina = 'csjnargentina',
  TCgermany = 'tcaleman',
  CourtRS = 'corteuk',
  CourtUS = 'supremecourtusa',
  CourtAustralia = 'hcaustralia',
  CCItaly = 'ccitaliana',
  Ccj = 'ccj_cursos',
  Ohchr = 'ohchr',
  Ejecutoria = 'ejecutoria',
  Coidh = 'coidh',
  CaseLawHp = 'caselawhp',
  CourtListener = 'courtlistener',
  Pjcr = 'pjcr',
  Telectoral = 'telectoral',
  Cadh = 'cadh',
  Cpi = 'cpi',
  CjcardMX = 'cjcardmx',
  Legcardmx = 'legcardmx'
}